refactor: Flatten CollisionFlagMap while optimizing
JavaScript handles memory differently than Kotlin so we need to design around the language.
There was a huge slowdown in this area of code so it's been simplified to work on the other components.

Disabled zone allocations and (for now) we're doing a single fixed alloc, which is only a reasonable option since we don't have to reserve space for instancing. Plan to be reworked.