fix(engine): add safety check to savefile loading
If for some reason we try to access savefile for a player that should
have one, and it's either absent from the filesystem or has invalid
data, we now reject the login attempt under a generic "Please try again"
message.

Previously, there were no sanity checks on reading the savefile data, or
if we didn't find a savefile for a player which should have one
(`logout_time != null`). In theory, if something went wrong here, it
would be sent to the player as a completely reset tutorial island
account. Then, when they logged out, the new savefile would be written.