fix(engine): input tracking bugfix
Current code has a bug where it does not properly wait for the client
data to be sent after tracking ends. It was supposed to wait a number of
ticks for the client to send this. As a result, we are often discarding
some input tracking data from the client.

```
if (this.waitingForLastReport) {
    if (this.endTrackingAt + InputTracking.FINAL_REPORT_TIME_LEEWAY >
World.currentTick) {
        // Finalize tracking
    }
}
```

This will not wait the intended duration before finalizing the tracking
session as the logic is wrong. `endTrackingAt+16` will always be greater
than `World.currentTick`. So if a client took more than 1 tick to send
its input tracking bundle it would be discarded.

Issue can be reproduced by delaying the FinishTracking ClientProt, or
modifying client to wait an extra tick before sending its input data.

In fixing this I have also reworked the class for better readability.