feat(accounts): upgrade to members account on members world login
This commit changes the default `members` setting for new accounts, *AND ALL EXISTING ACCOUNTS* to
`0` (non-members).

It at the same time adds a new config envvar, defaulting to true, that will automatically upgrade
accounts to members when they login to a `NODE_MEMBERS=1` world for the first time.

This is done as a way of keeping track of which accounts have remained truly f2p
post-members-launch. In the long term, we want to also add a page to the account centre that will
allow people to "lock in" their account as f2p-only, preventing members login entirely. For now
however, this should do the trick.

Without this, it will be impossible to verify which 'f2p-only' accounts have actually remained
f2p post-members-launch. In the f2p community, this will probably be a big deal.